change 2

+++ src/main/java/dao/AccountDao.java
@@ -4,7 +4,7 @@ import model.Account;
 
 public interface AccountDao {
 
-    String deposit(String accountNumber, int sum);
+    String deposit(String accountNumber, long sum);
 
     Account get(String accountNumber);
 }
